+++ .gitignore
@@ -475,3 +475,4 @@ FodyWeavers.xsd
 # Additional files built by Visual Studio
 
 # End of https://www.toptal.com/developers/gitignore/api/c,visualstudio,c++
+/Log.txt
