--- BaseCgr.cpp
@@ -1,4 +1,0 @@
-#include "BaseCgr.h"
-
-BaseCgr::~BaseCgr() {
-}
