--- GameCgr.cpp
@@ -1,5 +1,0 @@
-#include "GameCgr.h"
-
-GameCgr::~GameCgr() {
-
-}
