+++ header/Graphics.h
@@ -1,5 +1,5 @@
 #pragma once
-#include "Task.h"
+#include "..\header\Task.h"
 
 class Graphics : public Task
 {
