+++ source/BaseCgr.cpp
@@ -1,0 +1,4 @@
+#include "..\header\BaseCgr.h"
+
+BaseCgr::~BaseCgr() {
+}
