+++ source/BaseScene.cpp
@@ -1,4 +1,4 @@
-#include "BaseScene.h"
+#include "..\header\..\header\BaseScene.h"
 
 BaseScene::BaseScene(BaseCgr* changer) 
 	: baseCgr(changer){
