+++ source/FpsControll.cpp
@@ -1,4 +1,4 @@
-#include "FpsControll.h"
+#include "..\header\FpsControll.h"
 
 static int mStartTime;      //測定開始時刻
 static int mCount;          //カウンタ
