+++ source/GameCgr.cpp
@@ -1,0 +1,5 @@
+#include "..\header\GameCgr.h"
+
+GameCgr::~GameCgr() {
+
+}
