+++ source/GameScene.cpp
@@ -1,4 +1,4 @@
-#include "GameScene.h"
+#include "..\header\GameScene.h"
 
 GameScene::GameScene(GameCgr* changer,GameVar* gVar)
 	: gameCgr(changer)
