+++ source/GameVar.cpp
@@ -1,5 +1,5 @@
-#include "GameVar.h"
-#include "Dxlib.h"
+#include "..\header\GameVar.h"
+#include "..\dxlib_for_visual_studio\DxLib.h"
 
 GameVar::GameVar() {
 	//backGround = Graphics();
